Moved normalisations to its own folder
